fix: newly added then deleted node should not be exported (#576)
When we add nodes at export time, we filter based on nodes not having a
position on previous files, but we did not check the removal status,
resulting in added then removed nodes being exported.

It also change the delete API as using a payload with DELETE method does
not seems a good REST practice. The front-end code was amended
accordingly.

This PR also improve the tests with more reliable tests.

We also fixed OpenAPI specs and generated SDK (though we manually make
changes in SDK, because automatic generation did not give good results)

Fixes: #561